Upgrade LWJGL version 2 to 3
Because of a link posted in the discussion for #125, I suspect that it
may be the version of LWJGL that is causing the incorrect color channel
display. Specifically, the link to the same issue in Minecraft. My
theory is that the LWJGL version Minecraft shipped with was upgraded,
and as such, the bug does not appear in later versions.

Potentially resolves #125